Move code
from ApiConfigManagementService to LedgerConfigProvider
